Implement the buttons reactivity and output message.
Started the Vue.js lesson 03. Implemented the buttons and the correct
message output. However, I created a data to display the message and
a method to change the data variable but it didn't work. When I changed
the html to display the returned method, it worked. It would be good
to know why.