Fix the count reset watcher. COMPLETED.
I misunderstood the counter dynamics. It should reset the calculated
value to zero when the Result sentence change, not when the value is
re-calculated.
The purpose is to reset the counter to "Play" again, not to set a timer
to press the button like a speed game.
There is still room for improvemente about the buttons using the same
method.