Implement the user's input as the element's class.
I created a new data to hold the user's input. Then, I used the
third syntax explained in the video: I passed an array to the binded
class attribute in the html. Inside the array, there was the computed
property that returns the object and a binded data with the users input.
It worked like magic. I loved it!