Implement the visibility toggle button.
I implemented the button to toggle the visibility of the list by
adding a listener to the click, which activates a method that changes
a data property with a boolean expression. The list holds a v-if
statement that checks the data property with the boolean expression.
I added a <p> element with a v-else statement as a placeholder to keep
the two buttons apart and the list looks more organized.