Implement the game over screen.
I implemented the game over screen. The health are being watched by
watchers. And the whole game over screen is being triggered by a
v-if that checks for the winner data property. Initially, the winner
property is set to null, which is a falsy value. When one of the
watchers is triggered, the winner property is changed and it triggers
the v-ifs on the h3 html elements.