Add the main app and the component. Functional, but no style.
I created a main App.vue that contains two elements: <template> and
<script>
The template element contains the html tags that are going to be
inserted in the index.html with the Vue control. The script contains
the JS VUE code. The only difference is that the vue object now is
being extracted with the extract key word by defaul.
The component follows the same idea, but it is a convention that all
the components should be inside a folder called components.
Both main App.vue and the components are being imported in the main.js
file using the import keyword.
It's suggested to use the same file name to name the imported code
in the import code line.
After that, everything follow as usual. The main app is created using
the createApp() function and mounted. The component is 'mounted' using
the App.component() function.
Don't forget the App is a const related to the main app and in the
App.component() function gets two arguments:
The first is a two-words dash separated term, that doens't conflict
with an actual html tag and the secont argument is the imported
component.