Make the components reusable by using props.
The idea is to make a contact list with different contacts. And
creating a component to each contact is impractical. Therefore,
the component can be reusable by assigning data from to the component.
This can be achieved by adding the data inside the html tag respective
to the component from the App.vue. Then, all the data is refered in the
component without using this. or props. Plus, to help the component
understand what are those props, is created a props property, just
like data or methods, and describe the name of the properties.
VERY IMPORTANT: the properties in the html tags must be kebab-changes
<phone-number> and inside the component, it must be camelCase because
kebab-case is a invalid syntax for properties in JS.
The props property can be set as an array, which I did.