Implement the favorite toggle button - FIXED.
I implemented the favorite button by creating a new data property inside
the FriendContact component and assigning it to the props inherited from
App.vue.
Then, the button triggers the method that change the data property
inside the component. The value originally in the main App.vue
doesn't change. But the web app works as expected.