Implemented the #emit() to child -> parent data transfer.
The this.$emit() build-in JS function sends an event that can be
listened in the parent app. It can take many argument, being the first
a string with the name of the event (kebab-case suggested) and all
other arguments are data that will be passed to the method in the
parent element.

In the parent element, the event can be listened by v-on or @
and it can trigger a method. In this case, the method changes the
favorite property inside the friend object.
Now, the app works as expected changing the data inside the parent App,
which is the ideal.