Implement the add contact section.
There are couple inputs for the contact's info that v-model to the
object stored in the component's data object. Then, the button
unshift to the main App.vue the object. But it still buggy because
after adding the contact, the inputs are not reseted and every change
in the input change the contact info after being added.