Export and mount App and components.
I created the basic template/script tags on components and main App,
exported by default all three files and imported them in main.js.
I put a dummy text to see if the import/export flow is working.
NOTE - the main app.mount() must be placed at the end of the main.js,
or at least, after all app.components. If not, the components are not
rendered in the page.