Style improved. Need to fix bug - inputs don't clear.
I used the style tag inside each component to style the app and it is
looking better. There is room for improvement in the font and little
tweaks in the styling.
There is a bug that needs to be fixed. The inputs, after the submit,
keep the old data. Them don't reset.