Removed the topis props and replace it with provide/inject.
A better way to implement the logic is create a provide element inject
the vue object. The provide functions similarly to data.
It is better used declared as a function that returns an object that
references the data in the app.
After that, the topics object is being retrieved using the inject
keyword that is similar to props.