Change the emit system to provide/inject for the buttons.
The button in each item use to create an emit event that was
being passed through all parent elements until the top App.vue.
Now, I change that to the inject/provide system. The parent provides
the method through provide and the child has an event listener
that triggers the evend being passed in the inject array.