Importing components locally.
I imported the components BadgeList, TheHeader and UserInfo locally
inside the App.vue file. It's good practice to import your components
where they are being used. If a component is being used in different
files, then it should be global. If the component is being used in
one place only, it should be locally.
I prefer to use the following syntax for locally components:
components: {
    ImportedComponent
}
<ImportedComponent />