Add dynamic url (router)
I added/binded a url to the router in the ./teams page.
Fist, I created the binding route in the routers object in main.js.
Then, the route trigggers the component TeamMembers.
The TeamMembers component injects the user and team array from the App.
The TeamMembers component has an lifecycle hook (created) that initiate
a function to render the respective team members.
With different url after ./team , the created hook grasps the url with
the this.$route.params.teamId (the teamId is the binded term in routes).
Then, the hook loops through the teams and members arrays to return the
correct team and members to store it in the data() and then renders the
component.