Create dynamic links with :to=""
I implemented the links in each team (inside the teams page) to pass
the id through props and the team.id (each unity for the v-for by teams)
and using a computed property, I assigned the id to the url in the
component.
So, it is more scallable and I didn't have to hardcode all links.