Change the team loading to get from props received by props
Catching the t1/t2/t3 through $router is ok, but a better way is to
pass the url by props and it can be done setting the props form the url.
It works better and don't do that bug because there is always a correct
url being passed to the loadingTeamMembers.vue either from dynamic url
or fixed router-link.