Add two child in a transition element/ Mode property out-in
I added two buttons that are direct child of a transition element.
This can happen because both element don't exist in the DOM at the
same time. This is clear to Vue because I used v-else.
Also, I used the mode parameter in the transition to have more control
about the sequence of the animation.